multi_parrot_web/requirements.txt
Django==2.2.7
djangorestframework==3.9.4
django-cors-headers==3.0.2
django-cors-middleware==1.3.1